feature: Move other modules to wvlet.ai.core
