+++ ai-core/src/main/scala/wvlet/ai/core/control/ThreadUtil.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.ai.util.control
+package wvlet.ai.core.control
 
 import java.util.concurrent.ThreadFactory
 import java.util.concurrent.atomic.AtomicInteger
