+++ ai-core/src/main/scala/wvlet/ai/core/design/Binder.scala
@@ -16,7 +16,7 @@ package wvlet.ai.core.design
 import wvlet.ai.core.log.LogSupport
 import wvlet.ai.core.surface.Surface
 import LifeCycleHookType.*
-import wvlet.ai.util.SourceCode
+import wvlet.ai.core.util.SourceCode
 
 object Binder:
   sealed trait Binding extends Serializable:
