+++ ai-core/src/main/scala/wvlet/ai/core/design/Session.scala
@@ -14,7 +14,7 @@
 package wvlet.ai.core.design
 
 import wvlet.ai.core.surface.Surface
-import wvlet.ai.util.SourceCode
+import wvlet.ai.core.util.SourceCode
 
 /**
   * Session manages injected objects (e.g., Singleton)
